% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochar_field.R
\name{bc_dose_index}
\alias{bc_dose_index}
\title{Index of the Best Dose Option at a Carbon Price}
\usage{
bc_dose_index(ft, c_price)
}
\arguments{
\item{ft}{Output of \code{\link[=biochar_field_table]{biochar_field_table()}}.}

\item{c_price}{Carbon price (US$ Mg-1 CO2e).}
}
\value{
Integer vector (per cell) of the column of \code{ft} with the highest field value.
}
\description{
Index of the Best Dose Option at a Carbon Price
}
\keyword{internal}
