% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocharAG-package.R
\docType{package}
\name{BiocharAG-package}
\alias{BiocharAG}
\alias{BiocharAG-package}
\title{BiocharAG: Spatial Techno-Economic Model of Crop-Residue Bioenergy, BECCS and PyCCS}
\description{
Implements the C-SCAPE model. Compiled code: the cohort loop of the biochar field-application model
(\code{\link[=field_cohort_sums_cpp]{field_cohort_sums_cpp()}}, src/field_cohorts.cpp).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
