% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{calculate_annuity_factor}
\alias{calculate_annuity_factor}
\title{Calculate Annuity Factor}
\usage{
calculate_annuity_factor(discount_rate, lifetime)
}
\arguments{
\item{discount_rate}{Discount rate (decimal).}

\item{lifetime}{Lifetime in years.}
}
\value{
The annuity factor.
}
\description{
Calculates the Present Value of an Annuity Factor.
}
