% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bebcs.R
\name{calculate_bebcs_mode}
\alias{calculate_bebcs_mode}
\title{BEBCS Metrics for One Energy Mode}
\usage{
calculate_bebcs_mode(params)
}
\arguments{
\item{params}{A list of parameters with \code{bebcs_energy_mode} "power", "none" or "heat".}
}
\value{
A list of calculated metrics for BEBCS.
}
\description{
BEBCS Metrics for One Energy Mode
}
\keyword{internal}
