% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes.R
\name{calculate_bes_mode}
\alias{calculate_bes_mode}
\title{BES Metrics for One Operating Mode}
\usage{
calculate_bes_mode(params)
}
\arguments{
\item{params}{A list of parameters (\code{capacity_factor_bes}, \code{avg_dist_BES} and \code{elec_price} set for
the mode by \code{calculate_bes()}).}
}
\value{
A list of calculated metrics for BES.
}
\description{
BES Metrics for One Operating Mode
}
\keyword{internal}
