% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanence.R
\name{calculate_fperm_vectorized}
\alias{calculate_fperm_vectorized}
\title{Calculate Fperm from Pre-calculated Vector (Optimization)}
\usage{
calculate_fperm_vectorized(prep, soil_temp)
}
\arguments{
\item{prep}{A list object returned by \code{prep_fperm_vector}.}

\item{soil_temp}{Numeric vector of soil temperatures (Celsius).}
}
\value{
Numeric vector of Fperm fractions.
}
\description{
Calculates Fperm for a vector of soil temperatures using 1D linear interpolation
on a pre-calculated Fperm vector (created by \code{prep_fperm_vector}).
}
