% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{calculate_npv}
\alias{calculate_npv}
\title{Calculate Net Present Value}
\usage{
calculate_npv(cash_flows, discount_rate)
}
\arguments{
\item{cash_flows}{Vector of cash flows.}

\item{discount_rate}{Discount rate.}
}
\value{
NPV.
}
\description{
Calculate Net Present Value
}
