% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{calculate_rpv}
\alias{calculate_rpv}
\title{Calculate Relative Present Value (RPV)}
\usage{
calculate_rpv(results_list)
}
\arguments{
\item{results_list}{List of result objects from calculate_bes, calculate_beccs, calculate_bebcs.}
}
\value{
Data frame with comparison.
}
\description{
Calculate Relative Present Value (RPV)
}
