% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Default Parameters}
\usage{
default_parameters(scenario = list())
}
\arguments{
\item{scenario}{A named list of parameters to override the defaults. Defaults to an empty list.}
}
\value{
A named list of parameters.
}
\description{
Returns a list of default parameters used in the BiocharAG model.
Values inferred from op_space_2.41.xlsm.
}
