% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_regional_overrides}
\alias{get_regional_overrides}
\title{Get Regional Overrides from CSV}
\usage{
get_regional_overrides()
}
\value{
A named list of regional overrides.
}
\description{
Get Regional Overrides from CSV
}
