% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochar_valuation.R
\name{lime_value}
\alias{lime_value}
\title{Liming Value of an Acid-Neutralising Capacity}
\usage{
lime_value(params, anc)
}
\arguments{
\item{params}{Parameter list.}

\item{anc}{Acid-neutralising capacity (kg CaCO3-eq per Mg feed).}
}
\value{
Value in $/Mg feed.
}
\description{
Credited at the lime price (per Mg CaCO3-eq, delivered and spread) where soil pH is below \code{target_ph}.
}
\keyword{internal}
