% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{normalize_region_name}
\alias{normalize_region_name}
\title{Normalize Region Name}
\usage{
normalize_region_name(region)
}
\arguments{
\item{region}{Character string indicating region.}
}
\value{
Normalized region character string ("US", "India", "China", "Europe").
}
\description{
Standardizes regional names and aliases to match \code{regional_overrides} keys.
}
