% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_india.R
\name{parameters_india}
\alias{parameters_india}
\title{Default Parameters for India (North-West)}
\usage{
parameters_india()
}
\value{
A named list of parameters.
}
\description{
Deprecated: Returns a list of parameters customized for the Indian context
by calling \code{set_scenario(region = "India")}.
}
