% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochar_field.R
\name{pick_bc_dose}
\alias{pick_bc_dose}
\title{Field Effects of Given Dose Options}
\usage{
pick_bc_dose(ft, i)
}
\arguments{
\item{ft}{Output of \code{\link[=biochar_field_table]{biochar_field_table()}}.}

\item{i}{Dose option index per cell (or a single index for all cells).}
}
\value{
As \code{\link[=choose_bc_dose]{choose_bc_dose()}}.
}
\description{
Field Effects of Given Dose Options
}
\keyword{internal}
