% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sampling.R
\name{qpert}
\alias{qpert}
\title{Quantile Function of the (Modified) PERT Distribution}
\usage{
qpert(p, min, mode, max, lambda = 4)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{min, mode, max}{Lower bound, most likely value and upper bound.}

\item{lambda}{Shape parameter (4 = classic PERT).}
}
\value{
Vector of quantiles.
}
\description{
Quantile Function of the (Modified) PERT Distribution
}
