% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_sweep.R
\name{quad_roots01}
\alias{quad_roots01}
\title{Roots of c0 + b w + a w^2 in \link{0, wmax}}
\usage{
quad_roots01(c0, b, a, wmax)
}
\arguments{
\item{c0, b, a}{Numeric vectors of equal length.}

\item{wmax}{Upper bound of w (scalar, may be Inf).}
}
\value{
A two-column matrix of the roots in \link{0, wmax} (smaller first; NA where absent).
}
\description{
Roots of c0 + b w + a w^2 in \link{0, wmax}
}
\keyword{internal}
