% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{regional_overrides}
\alias{regional_overrides}
\title{Regional Overrides List}
\format{
An object of class \code{list} of length 4.
}
\usage{
regional_overrides
}
\description{
A predefined list of regional parameter overrides for non-spatial parameters
(financial, capital cost modifiers, O&M labor factors, and fertilizer prices).
}
\keyword{datasets}
