% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{resolve_plant_mw_th}
\alias{resolve_plant_mw_th}
\title{Resolve plant_mw_th for a specific technology}
\usage{
resolve_plant_mw_th(plant_mw_th, tech)
}
\arguments{
\item{plant_mw_th}{A single numeric value or named vector of numeric values.}

\item{tech}{Character string ("BES", "BECCS", or "BEBCS").}
}
\value{
A single numeric value.
}
\description{
Resolve plant_mw_th for a specific technology
}
