% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{scenarios_base}
\alias{scenarios_base}
\title{Scenarios List}
\format{
An object of class \code{list} of length 8.
}
\usage{
scenarios_base
}
\description{
A predefined list of scenarios used to override default parameters.
Default plant size is based on https://www.pfpi.net/biomass-basics which gives an average biomass facility currently being built as 38 MWe. With an assumed electrical efficiency of 30\%, this gives a thermal capacity of 127 MWth, which we round to 125 MWth for simplicity.
}
\keyword{datasets}
