% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_scenario}
\alias{set_scenario}
\title{Set Scenario Parameters}
\usage{
set_scenario(scenario = list())
}
\arguments{
\item{scenario}{A named list of parameters to override the defaults. Defaults to an empty list.}
}
\value{
A named list of parameters.
}
\description{
Returns a list of default parameters overridden by a specific scenario.
Values inferred from op_space_2.41.xlsm and tea_literature_review.md.
}
