% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_sweep.R
\name{sweep_abate}
\alias{sweep_abate}
\title{Abatement at a Carbon Price, Interpolated from a Sweep}
\usage{
sweep_abate(sweep, cp)
}
\arguments{
\item{sweep}{Output of \code{run_price_sweep}.}

\item{cp}{Carbon price ($/tCO2), a scalar.}
}
\value{
Cells x 3 matrix of abatement (tCO2e/Mg).
}
\description{
Abatement at a Carbon Price, Interpolated from a Sweep
}
