% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_sweep.R
\name{sweep_net}
\alias{sweep_net}
\title{Net Value at a Carbon Price, Interpolated from a Sweep}
\usage{
sweep_net(sweep, cp)
}
\arguments{
\item{sweep}{Output of \code{run_price_sweep}.}

\item{cp}{Carbon price ($/tCO2), a scalar.}
}
\value{
Cells x 3 matrix of net value ($/Mg).
}
\description{
Net Value at a Carbon Price, Interpolated from a Sweep
}
