% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_sweep.R
\name{sweep_to_raster}
\alias{sweep_to_raster}
\title{Map Cell Vectors from a Sweep onto a Raster}
\usage{
sweep_to_raster(sweep, m, names = colnames(m))
}
\arguments{
\item{sweep}{Output of \code{run_price_sweep}.}

\item{m}{Vector or cells x k matrix.}

\item{names}{Layer names.}
}
\value{
SpatRaster on the sweep's template.
}
\description{
Map Cell Vectors from a Sweep onto a Raster
}
