% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bebcs.R
\name{calculate_bebcs}
\alias{calculate_bebcs}
\title{Calculate Biochar-Energy (BEBCS) Metrics}
\usage{
calculate_bebcs(params)
}
\arguments{
\item{params}{A list of parameters.}
}
\value{
A list of calculated metrics for BEBCS.
}
\description{
Calculate Biochar-Energy (BEBCS) Metrics
}
