% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beccs.R
\name{calculate_beccs}
\alias{calculate_beccs}
\title{Calculate Bioenergy Carbon Capture and Storage (BECCS) Metrics}
\usage{
calculate_beccs(params)
}
\arguments{
\item{params}{A list of parameters.}
}
\value{
A list of calculated metrics for BECCS.
}
\description{
Calculate Bioenergy Carbon Capture and Storage (BECCS) Metrics
}
