% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes.R
\name{calculate_bes}
\alias{calculate_bes}
\title{Calculate Bioenergy System (BES) Metrics}
\usage{
calculate_bes(params)
}
\arguments{
\item{params}{A list of parameters.}
}
\value{
A list of calculated metrics for BES.
}
\description{
Modernized logic (2024 Basis):
\itemize{
\item Uses modernized Capital Cost ($3,000/kW) and Efficiency (30\%) defaults.
\item Calculates Levelized Cost of Electricity components (CAPEX/OPEX).
\item Explicitly tracks Scope 3 transport emissions and road tortuosity.
}
}
