% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bes.R
\name{calculate_bes}
\alias{calculate_bes}
\title{Calculate Bioenergy System (BES) Metrics}
\usage{
calculate_bes(params)
}
\arguments{
\item{params}{A list of parameters.}
}
\value{
A list of calculated metrics for BES.
}
\description{
Calculate Bioenergy System (BES) Metrics
}
