% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default Parameters Dataset}
\format{
A named list.
}
\usage{
default_parameters
}
\description{
A list containing the default parameters for the BiocharAG model.
}
\keyword{datasets}
