% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Regionalized Scenarios}
\format{
An object of class \code{list} of length 12.
}
\usage{
scenarios
}
\description{
A predefined list of scenarios that includes global and regional parameter overrides.
}
\keyword{datasets}
