% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Scenarios List}
\format{
An object of class \code{list} of length 5.
}
\usage{
scenarios
}
\description{
A predefined list of scenarios used to override default parameters.
}
\keyword{datasets}
