% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{scenarios_base}
\alias{scenarios_base}
\title{Scenarios List}
\format{
An object of class \code{list} of length 6.
}
\usage{
scenarios_base
}
\description{
A predefined list of scenarios used to override default parameters.
}
\keyword{datasets}
